+++ android/src/main/java/com/ReactNativeBlobUtil/ReactNativeBlobUtilImpl.java
@@ -294,7 +294,7 @@ class ReactNativeBlobUtilImpl {
         }
     }
 
-    public void slice(String src, String dest, int start, int end, Promise promise) {
+    public void slice(String src, String dest, long start, long end, Promise promise) {
         ReactNativeBlobUtilFS.slice(src, dest, start, end, "", promise);
     }
